+++ internal/crossdock/client/echo/thrift.go
@@ -27,7 +27,7 @@ import (
 	disp "go.uber.org/yarpc/internal/crossdock/client/dispatcher"
 	"go.uber.org/yarpc/internal/crossdock/client/random"
 	"go.uber.org/yarpc/internal/crossdock/thrift/echo"
-	"go.uber.org/yarpc/internal/crossdock/thrift/echo/yarpc/echoclient"
+	"go.uber.org/yarpc/internal/crossdock/thrift/echo/echoclient"
 
 	"github.com/crossdock/crossdock-go"
 )